+++ src/plugins/ncr/services/SluggableSubscriber.js
@@ -2,7 +2,6 @@ import { change, formValueSelector } from 'redux-form';
 import addDateToSlug from '@gdbots/common/addDateToSlug';
 import createSlug from '@gdbots/common/createSlug';
 import EventSubscriber from '@gdbots/pbjx/EventSubscriber';
-import isValidSlug from '@gdbots/common/isValidSlug';
 import slugContainsDate from '@gdbots/common/slugContainsDate';
 
 export default class SluggableSubscriber extends EventSubscriber {
@@ -11,7 +10,6 @@ export default class SluggableSubscriber extends EventSubscriber {
     this.onInitForm = this.onInitForm.bind(this);
     this.onSubmitForm = this.onSubmitForm.bind(this);
     this.onTitleChanged = this.onTitleChanged.bind(this);
-    this.onValidateForm = this.onValidateForm.bind(this);
     this.sluggableForms = sluggableForms;
   }
 
@@ -81,47 +79,6 @@ export default class SluggableSubscriber extends EventSubscriber {
     store.dispatch(change(form, 'slug', slug));
   }
 
-  /**
-   * This if the redux form sync validation process.
-   *
-   * @param {FormEvent} formEvent
-   */
-  onValidateForm(formEvent) {
-    const node = formEvent.getMessage();
-    const { slug } = formEvent.getData();
-    const isDatedSlug = this.shouldUseDatedSlug(formEvent.getProps().form);
-
-    if (slug) {
-      // slug format check
-      if (!isValidSlug(slug, isDatedSlug)) {
-        let error = '';
-        if (isDatedSlug) {
-          error = 'Invalid slug format - this slug cannot start or end with a dash or slash and can only contain lowercase alphanumeric characters, dashes, or slashes.';
-        } else {
-          error = 'Invalid slug format - this slug cannot start or end with a dash and can only contain lowercase alphanumeric characters or dashes.';
-        }
-        formEvent.addError('slug', error);
-        return;
-      }
-
-      // check if slug contains date, only if datedSlug config is set to true
-      if (isDatedSlug && !slugContainsDate(slug)) {
-        formEvent.addError('slug', 'Slug must have a date in YYYY/MM/DD format.');
-        return;
-      }
-
-      try {
-        node.set('slug', slug);
-      } catch (e) {
-        formEvent.addError('slug', e.message);
-      }
-    }
-
-    if (!slug) {
-      formEvent.addError('slug', 'Please enter a valid slug.');
-    }
-  }
-
   shouldUseDatedSlug(formName) {
     return this.sluggableForms[formName] || this.sluggableForms.default;
   }
@@ -130,7 +87,6 @@ export default class SluggableSubscriber extends EventSubscriber {
     return {
       '@@redux-form/BLUR': this.onTitleChanged,
       'gdbots:ncr:mixin:sluggable.init_form': this.onInitForm,
-      'gdbots:ncr:mixin:sluggable.validate_form': this.onValidateForm,
       'gdbots:ncr:mixin:sluggable.submit_form': this.onSubmitForm,
     };
   }